# swagger/swagger.yaml
swagger: "2.0"
info:
  description: "This is an API for a simple online market. "
  version: "1.0.0"
  title: "Online Marketplace"
host: "localhost:8000"
basePath: "/api"
...
schemes:
- "http"
paths:
...
  /order:
    post:
      tags:
      - "order"
      summary: "Create an order"
      operationId: "addOrder"
      consumes:
      - "application/json"
      
      parameters:
      - in: "body"
        name: "body"
        description: "New order object"
        required: true
        schema:
           type: "object"
           properties:
              firstName:
                type: "string"
                example: "Kevin"
              lastName:
                type: "string"
                example: "Owens"
              products:
                type: "array"
                items: 
                    type: "object"
                    properties:
                      productId: 
                        type: integer
                        example: 2
                      quantity:
                        type: "integer"
                        example: 3
      responses:
        201:
          description: "Successful operation. The order has been successfully created."
        400:
          description: "Bad request. Invalid Parameters"
...